pyproject.toml
[tool.poetry]
name = "pa-backend"
version = "0.1.0"
description = "AI Gym Tracker + Personal Assistant backend"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115"
uvicorn = {version = "^0.34", extras = ["standard"]}
sqlalchemy = {version = "^2.0", extras = ["asyncio"]}
asyncpg = "^0.30"
alembic = "^1.14"
pydantic-settings = "^2.7"
pgvector = "^0.3"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.22"
email-validator = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
